requirements.txt
streamlit==1.29.0
langchain==0.0.286
langchain-community==0.0.16
langchain-huggingface==0.0.1
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
PyPDF2==3.0.1
graphviz==0.20.1
plotly==5.18.0
psutil==5.9.6
matplotlib==3.8.2
langgraph==0.0.20
pillow==10.1.0
pypdf==3.17.1
chromadb==0.4.18
